tuning the model

+++ CSCE636-project-2022/starter_code/DataLoader.py
@@ -37,7 +37,7 @@ def load_data(data_dir):
     y_train = np.array(y_train)
     with open(os.path.join(data_dir, 'test_batch'), 'rb') as fo:
         data = pickle.load(fo, encoding='bytes')
-        x_test = data[b'data'].astype(np.int64)
+        x_test = data[b'data'].astype(np.uint8)
         y_test = np.array(data[b'labels'])
 
     ### END CODE HERE
